pyproject.toml
[tool.poetry]
name = "anmitsu"
version = "0.1.0"
description = "Azuki Watcher"
authors = ["noaione <noaione0809@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.3"
APScheduler = "^3.9.1"
aiosqlite = "^0.17.0"
PyYAML = "^6.0"
orjson = "^3.8.0"
torf = "^4.0.4"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
